pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autoinvoice"
version = "2.0.0"
description = "CARICOM invoice processing pipeline - PDF to XLSX customs declarations"
requires-python = ">=3.11"
dependencies = [
    "pdfplumber>=0.10",
    "PyMuPDF>=1.23",
    "pytesseract>=0.3",
    "Pillow>=10",
    "openpyxl>=3.1",
    "pyyaml>=6",
    "pydantic>=2",
    "pydantic-settings>=2",
    "anthropic>=0.30",
    "httpx>=0.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-bdd>=7",
    "pytest-cov>=5",
    "pytest-xdist>=3",
    "ruff>=0.4",
    "mypy>=1.10",
    "pre-commit>=3",
]

[project.scripts]
autoinvoice = "autoinvoice.composition.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/autoinvoice"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--strict-markers --tb=short -q"
markers = [
    "slow: marks tests that take >5s",
    "integration: marks integration tests needing real files",
    "bdd: marks BDD acceptance tests",
]

[tool.coverage.run]
source = ["src/autoinvoice"]
omit = [
    "src/autoinvoice/composition/*",
    "src/autoinvoice/domain/ports/*",   # Protocol ABCs - no implementation to test
]

[tool.coverage.report]
fail_under = 70
show_missing = true
exclude_lines = [
    "if TYPE_CHECKING:",
    "pragma: no cover",
    "raise NotImplementedError",
    "@(abc\\.)?abstractmethod",
    "class .*Protocol.*:",
]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "C4", "SIM", "TCH"]

[tool.ruff.lint.isort]
known-first-party = ["autoinvoice"]

[tool.mypy]
strict = true
warn_return_any = true
warn_unused_configs = true
mypy_path = "src"
packages = ["autoinvoice"]

[[tool.mypy.overrides]]
module = [
    "fitz",
    "fitz.*",
    "pytesseract",
    "pytesseract.*",
    "pdfplumber",
    "pdfplumber.*",
    "PIL",
    "PIL.*",
]
ignore_missing_imports = true
